backend/requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.2
pydantic-settings>=2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0

# LangChain & LangGraph
langchain>=0.1.0
langgraph>=0.0.26
langchain-openai>=0.0.5
openai>=1.10.0
langsmith>=0.1.0

# Utils
python-multipart==0.0.6
httpx==0.25.2
redis==5.0.1
uuid==1.30
Pillow>=10.0.0

# File Processing
openpyxl>=3.1.2  # Excel file support
PyPDF2>=3.0.1  # PDF file support
python-docx>=1.1.0  # Word document support

# Vector Database
pgvector==0.3.6
numpy>=1.24.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
isort==5.13.2